pyproject.toml
[project]
name = "nova-retrieve"
version = "0.1.0"
description = "Enterprise-grade Agentic RAG built on LangChain + LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.7",
    "langchain-core>=0.3.15",
    "langchain-community>=0.3.5",
    "langchain-openai>=0.2.5",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.2",
    "langchain-huggingface>=0.1.2",
    "langgraph>=0.2.50",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.1",
    "tavily-python>=0.5.0",
    "fastapi>=0.115.4",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.3",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "python-multipart>=0.0.17",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "unstructured[md]>=0.16.5",
    "beautifulsoup4>=4.12.3",
    "loguru>=0.7.2",
    "httpx>=0.27.2",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"
